Refactor init config
