+++ app/controllers/general.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once __DIR__ . '/../init.php';
+require_once __DIR__ . '/../init/http.php';
 
 use Ahc\Jwt\JWT;
 use Appwrite\Auth\Auth;
