+++ tests/unit/Event/EventTest.php
@@ -11,7 +11,7 @@ use Utopia\Queue;
 use Utopia\Queue\Client;
 use Utopia\System\System;
 
-require_once __DIR__ . '/../../../app/init.php';
+require_once __DIR__ . '/../../../app/init/http.php';
 
 class EventTest extends TestCase
 {
